init send comando

+++ RemoteBash/app/src/main/java/com/remotebash/model/UsuarioModel.kt
@@ -1,12 +1,13 @@
 package com.remotebash.model
 
 class UsuarioModel() {
+
     constructor(email:String, password:String) : this() {
         this.email = email
         this.password = password
     }
 
-    var id:Long? = null
+    var id:Int? = null
     var name:String? = null
     var password:String? = null
     var cellphone:String? = null
